fix(dshot): fit the programming guards in the F051 size budget
Main grew after this branch was cut, and the instrumented F051 build landed 12 bytes over the size gate. Ending the stage on every frame and advancing only on a valid one drops the separate abort paths, and counting the deadline in the 1 kHz block costs 16 bytes less than the 20 kHz tick. That block is skipped during sine start, so the abort check uses escIsDriving(): a sine start now aborts programming the same way running does.

+++ Inc/dshot.h
@@ -96,8 +96,8 @@ enum {
 void computeDshotDMA(void);
 void make_dshot_package(uint16_t com_time);
 
-/* Decremented by the control tick; only the decoder starts a transaction. */
-extern volatile uint16_t dshot_programming_ticks;
+/* Decremented by the 1 kHz control tick; only the decoder starts a transaction. */
+extern volatile uint8_t dshot_programming_ticks;
 
 extern void playInputTune(void);
 extern void playInputTune2(void);
